Revert "View games ui"
