+++ resources/assets/js/mainApp/components/scheduledGames/ScheduleGames.js
@@ -1,8 +1,8 @@
 import React, { Component, Fragment } from 'react'
-import GameFilter from './GameFilter'
-import GameList from './GameList'
-import GameDetails from './GameDetails'
-import SingleGameDetails from './SingleGameDetails.js'
+import GameFilter from './gameFilter'
+import GameList from './gameList'
+import GameDetails from './gameDetails'
+import SingleGameDetails from './singlegameDetails'
 import { PullDataFunction as getScheduleGames } from './getScheduleGames'
 import axios from 'axios'
 
@@ -21,7 +21,6 @@ export default class ScheduleGames extends Component {
       singleView: false,
       moreplease: true,
       counter: 1,
-      scheduleGamesView: {},
     }
   }
 
@@ -31,7 +30,7 @@ export default class ScheduleGames extends Component {
     if (id) {
       const scheduleGames = await axios.get(`/api/ScheduleGame/filtered_by_one/${id}`)
       if (scheduleGames.data && scheduleGames.data.latestScheduledGames.length > 0) {
-        this.setState({ scheduleGamesView: scheduleGames.data, showRightSideInfo: true, singleView: true })
+        this.setState({ scheduleGames: scheduleGames.data, showRightSideInfo: true, singleView: true })
       }
     } else {
       const scheduleGames = await getScheduleGames({ counter: 1 })
@@ -53,7 +52,6 @@ export default class ScheduleGames extends Component {
   }
 
   handleChange = async (data, name) => {
-    this.setState({ singleScheduleGamesPayload: {}, showRightSideInfo: false })
     if (name == 'game_name') {
       this.setState({ ...data }, () => {
         this.getScheduleGamesChangeCall()
@@ -66,6 +64,7 @@ export default class ScheduleGames extends Component {
   }
   getScheduleGamesChangeCall = async (data = {}) => {
     const { counter, scheduleGames = [] } = this.state
+    window.scrollTo(0, 0)
     const scheduleGamesRes = await getScheduleGames({ ...this.state, ...data, counter: 1 })
 
     if (scheduleGamesRes.data && scheduleGamesRes.data.latestScheduledGames.length == 0) {
@@ -125,9 +124,7 @@ export default class ScheduleGames extends Component {
       showRightSideInfo,
       commentData,
       singleView,
-      scheduleGamesView = {},
     } = this.state
-    const { latestScheduledGames = [] } = scheduleGamesView
 
     if (this.props.initialData == 'loading') {
       return <h1>Loading</h1>
@@ -135,7 +132,7 @@ export default class ScheduleGames extends Component {
     return (
       <section className='viewGame__container'>
         {id == '' && <GameFilter handleChange={this.handleChange} />}
-        <div className={`gameList__section ${singleView ? 'singleGameView__container' : 'GameView__container'}`}>
+        <div className={`gameList__section ${singleView ? 'singleGameView__container' : ''}`}>
           {!singleView && scheduleGames.length > 0 ? (
             <Fragment>
               <InfiniteScroll dataLength={scheduleGames.length} next={this.getScheduleGamesData} hasMore={this.state.moreplease}>
@@ -151,18 +148,12 @@ export default class ScheduleGames extends Component {
                 selected_game={selected_game}
                 showRightSideInfo={showRightSideInfo}
                 commentData={commentData}
-                {...this.props}
               />
             </Fragment>
           ) : (
             <Fragment>
-              {latestScheduledGames.length > 0 ? (
-                <SingleGameDetails
-                  scheduleGames={scheduleGamesView}
-                  showRightSideInfo={showRightSideInfo}
-                  commentData={commentData}
-                  {...this.props}
-                />
+              {scheduleGames.length > 0 ? (
+                <SingleGameDetails scheduleGames={scheduleGames} showRightSideInfo={showRightSideInfo} commentData={commentData} />
               ) : (
                 <h1>There is no data found!</h1>
               )}
