+++ resources/assets/js/mainApp/components/scheduledGames/gameDetails.js
@@ -1,32 +1,28 @@
 import React, { Component, Fragment } from 'react'
 import moment from 'moment'
-import Approved_gamers from './ApprovedGamers'
-import JoinButtonAction from './JoinButtonAction'
-import GameComments from './GameComments'
-
+const buttonStatus = {
+  '0': 'Join',
+  '1': 'Joined',
+  '3': 'Pending',
+}
 export default class GameDetails extends Component {
   constructor() {
     super()
     this.state = {}
   }
-  handleShowAllComments = () => {
-    this.setState({ showAllComment: !this.state.showAllComment })
-  }
 
   render() {
-    const { singleScheduleGamesPayload = {}, selected_game = {}, showRightSideInfo, commentData, fetching } = this.props
+    const { singleScheduleGamesPayload = {}, selected_game = {}, showRightSideInfo, commentData } = this.props
     const { additional_game_info = [], approved_gamers = [], join_status = '' } = singleScheduleGamesPayload
     const [game_additional_data = {}] = additional_game_info
-    const { id = '', game_name = '', experience = '', no_of_gamers = '', tags = [] } = selected_game
+    const { game_name = '', experience = '', no_of_gamers = '', tags = [] } = selected_game
     const { start_date_time = '', end_date_time = '', limit, description = '', platform = '', region = '' } = game_additional_data
 
     const { no_of_comments = [], lastComment = '' } = commentData
 
-    const { showAllComment } = this.state
-
     return (
       <div className='gameDetails'>
-        {showRightSideInfo && !showAllComment && (
+        {showRightSideInfo && (
           <Fragment>
             <div className='gameDetails__header'>
               <div className='gameName'>
@@ -43,36 +39,38 @@ export default class GameDetails extends Component {
                 </div>
                 {experience && <div className='game__level'>{experience}</div>}
               </div>
-
-              <JoinButtonAction join_status={join_status} />
+              {join_status == 0 ? (
+                <div className='game__action__buttton'>
+                  <button type='button'>{`Join`}</button>
+                </div>
+              ) : (
+                <div className='game__action__buttton'>
+                  <button type='button'>{buttonStatus[`${join_status}`]}</button>
+                </div>
+              )}
             </div>
             <div className='gameDetails__body'>
               <div className='filter__label'>Game Details</div>
-              {description && <div className='gameDescription'>Description</div>}
-              {description && <div className='gameDescription__body'>{description}</div>}
+              <div className='gameDescription'>Description</div>
+              <div className='gameDescription__body'>{description}</div>
               <div className='gameTime__label'>End Time</div>
               <div className='gameTime__value'>{moment(end_date_time).format('LLLL')}</div>
-              {tags.length > 0 && <div className='gameTags__label'>Tags</div>}
-
+              <div className='gameTags__label'>Tags</div>
+              <div className='gameTags__value'>
+                {tags &&
+                  tags.length > 0 &&
+                  tags.slice(0, 7).map((tag) => {
+                    return <div className='singleTags'>{tag.content}</div>
+                  })}
+              </div>
               {platform && <div className='gameTime__label'>Platform</div>}
               {platform && <div className='gameTime__value'>{platform}</div>}
               {region && <div className='gameTime__label'>Region</div>}
               {region && <div className='gameTime__value'>{region}</div>}
-              <Approved_gamers approved_gamers={approved_gamers} />
-              {tags.length > 0 && <div className='gameTags__label'>Tags</div>}
-              <div className='gameTags__value'>
-                {tags &&
-                  tags.length > 7 &&
-                  tags.map((tag) => {
-                    return <div className='singleTags'>{tag.content}</div>
-                  })}
-              </div>
             </div>
-            {(lastComment || !showAllComment) && (
+            {lastComment && (
               <div className='gameDetaiils__footer'>
-                <div className='view__all__comments' onClick={this.handleShowAllComments}>
-                  View all (3) comments
-                </div>
+                <div className='view__all__comments'>View all (3) comments</div>
                 <div className='game__comment'>
                   <div className='profile__image'></div>
                   <div className='arrow'></div>
@@ -85,14 +83,6 @@ export default class GameDetails extends Component {
               </div>
             )}
           </Fragment>
-        )}
-        {showAllComment && (
-          <GameComments
-            game_id={id}
-            scheduleGames_data={selected_game}
-            user={this.props.initialData}
-            toggleBack={this.handleShowAllComments}
-          />
         )}
       </div>
     )
