+++ resources/assets/js/mainApp/components/scheduledGames/singlegameDetails.js
@@ -1,23 +1,19 @@
 import React, { Component, Fragment } from 'react'
 import moment from 'moment'
-import Approved_gamers from './ApprovedGamers'
-import JoinButtonAction from './JoinButtonAction'
-import GameComments from './GameComments'
-
+const buttonStatus = {
+  '0': 'Join',
+  '1': 'Joined',
+  '3': 'Pending',
+}
 export default class GameDetails extends Component {
   constructor() {
     super()
-    this.state = {
-      showAllComment: false,
-    }
-  }
-  handleShowAllComments = () => {
-    this.setState({ showAllComment: !this.state.showAllComment })
+    this.state = {}
   }
 
   render() {
     const { scheduleGames = [], showRightSideInfo, commentData } = this.props
-    const { latestScheduledGames, approved_gamers = [], join_status } = scheduleGames
+    const { latestScheduledGames, approved_gamers, join_status } = scheduleGames
     const [scheduleGames_data = {}] = latestScheduledGames
     const {
       game_name = '',
@@ -30,15 +26,12 @@ export default class GameDetails extends Component {
       description = '',
       platform = '',
       region = '',
-      id = '',
     } = scheduleGames_data
     const { no_of_comments = [], lastComment = '' } = commentData
 
-    const { showAllComment } = this.state
-
     return (
       <div className='gameDetails'>
-        {showRightSideInfo && !showAllComment && (
+        {showRightSideInfo && (
           <Fragment>
             <div className='gameDetails__header'>
               <div className='gameName'>
@@ -55,7 +48,15 @@ export default class GameDetails extends Component {
                 </div>
                 {experience && <div className='game__level'>{experience}</div>}
               </div>
-              <JoinButtonAction join_status={join_status} />
+              {join_status == 0 ? (
+                <div className='game__action__buttton'>
+                  <button type='button'>{`Join`}</button>
+                </div>
+              ) : (
+                <div className='game__action__buttton'>
+                  <button type='button'>{buttonStatus[`${join_status}`]}</button>
+                </div>
+              )}
             </div>
             <div className='gameDetails__body'>
               <div className='filter__label'>Game Details</div>
@@ -67,7 +68,7 @@ export default class GameDetails extends Component {
               <div className='gameTags__value'>
                 {tags &&
                   tags.length > 0 &&
-                  tags.map((tag) => {
+                  tags.slice(0, 7).map((tag) => {
                     return <div className='singleTags'>{tag.content}</div>
                   })}
               </div>
@@ -75,13 +76,10 @@ export default class GameDetails extends Component {
               {platform && <div className='gameTime__value'>{platform}</div>}
               {region && <div className='gameTime__label'>Region</div>}
               {region && <div className='gameTime__value'>{region}</div>}
-              <Approved_gamers approved_gamers={approved_gamers} />
             </div>
-            {(lastComment || !showAllComment) && (
+            {lastComment && (
               <div className='gameDetaiils__footer'>
-                <div className='view__all__comments' onClick={this.handleShowAllComments}>
-                  View all (3) comments
-                </div>
+                <div className='view__all__comments'>View all (3) comments</div>
                 <div className='game__comment'>
                   <div className='profile__image'></div>
                   <div className='arrow'></div>
@@ -94,16 +92,6 @@ export default class GameDetails extends Component {
               </div>
             )}
           </Fragment>
-        )}
-        {showAllComment && (
-          <div className='gameDetails'>
-            <GameComments
-              game_id={id}
-              toggleBack={this.handleShowAllComments}
-              scheduleGames_data={scheduleGames_data}
-              user={this.props.initialData}
-            />
-          </div>
         )}
       </div>
     )
